requirements.txt
Flask
openai
matplotlib
python-dotenv
croniter
pandas
plotly
kaleido